+++ src/test/kotlin/racingcar/domain/CarsTest.kt
@@ -3,7 +3,7 @@ package racingcar.domain
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import racingcar.common.PowerStrategy
+import racingcar.domain.powerstrategy.PowerStrategy
 
 class CarsTest {
